requirements.txt
# Core dependencies
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Diffusion models
diffusers>=0.24.0
transformers>=4.35.0
accelerate>=0.24.0

# LoRA
peft>=0.6.0

# Optimization
xformers>=0.0.22
einops>=0.7.0

# Face processing
facenet-pytorch>=2.5.0

# API
fastapi>=0.104.0
uvicorn>=0.24.0
gradio>=4.0.0

# Audio
torchaudio>=2.1.0
librosa>=0.10.0

# Video
opencv-python>=4.8.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.9

# Data
numpy>=1.24.0
pillow>=10.0.0

# Config
pyyaml>=6.0.1
omegaconf>=2.3.0

# Utilities
tqdm>=4.66.0
safetensors>=0.4.0

# Optional: FP8 (Ada/Hopper only)
# transformer-engine>=1.0; platform_machine=='x86_64'

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
